OSIS-4526 business validator ignore success message to add
The reason why is that a BusinessListValidator could contain
a BusinessListValidtor as validator. In that case, an exception
was raised due to the fact that the BusinessListValidator
contained returned a success message on successfull validation.
For this reason, it was prefered to ignore success message when
adding than raise an exception.